fix: bugfix in level allocation
level_set_size() takes two parameters shape_in and
buffer_in, which were previously set explicitly, i.e.,
without a function call.

From how these two values are used, they seem to mean
the following:
- level%lev_shape = shape_in: a user defined array
containing general level info, e.g.,
but not necessarily only, the number of DOF
- level%mpibuflen = buffer_in: size of the MPI buffer that gets sent
in double precision. Previously, this was called nvars in LibPFASST.
In our case, this just corresponds to the number of DOF on one level

The previous definition lead to a bug for more than one level and a
high resolution, since the default for mpibuflen was product(buffer_in).
This lead to an integer overflow which was caught by the sanity check
for the MPI buffer length in pf_pfasst.f90

I updated the definitions of the SDC programs to be more clear.